requirements.txt
python-telegram-bot==21.6
stripe==11.1.0
openai==1.55.0
requests==2.32.3
python-dotenv==1.0.1
